+++ src/components/core/actions/index.ts
@@ -1,2 +1,3 @@
 export { default as ActionButton } from './ActionButton';
+export { default as FloatingButton } from './FloatingButton';
 export { default as IconButton } from './IconButton';