+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/), and t
 ### 🌱 Added
 
 - Add new core components
+  - FloatingButton (actions)
   - NavTable (data)
   - StatusLight (status)
 - Add new icons